Merging pull request for the inside-out refactoring of transactions and accounts
After using the Account and Transaction objects to model real UI I want to build it became clear
that the design was inside-out. Accounts holding transactions' list was not the right metaphor,
despite being an obvious one. What really happens is that we have a transactions' logbook, in which
transactions are entered. Each transaction contains one or more atomic operations. Each operation
pertains to a single account and update this account's balance by a given amount.

This pull request contains the refactoring from the original design to the new design.

Link to pull request: https://github.com/forloop7/beancounter/pull/1